fix: seek method did notify delegate to early
When seeking the jukebox didn't wait for AVPlayer to finish seeking before notifying delegate.
When delegate receives this message, progress is still on state before seeking.

AVPlayer offers a method with completionBlock, which is called when seeking is finished. 
Using this completionBlock to notify the delegate resolves some issues with seeking.